+++ src/main/java/vip/gadfly/chakkispring/dto/admin/NewClassDTO.java
@@ -19,5 +19,5 @@ public class NewClassDTO {
     private String info;
 
     @Min(value = 1, message = "{semester.id.positive}")
-    private Long semesterId;
+    private Integer semesterId;
 }
