+++ src/main/java/vip/gadfly/chakkispring/mapper/PermissionMapper.java
@@ -20,7 +20,7 @@ public interface PermissionMapper extends BaseMapper<PermissionDO> {
      * @param groupIds 分组ids
      * @return 权限
      */
-    List<PermissionDO> selectPermissionsByGroupIds(@Param("groupIds") List<Long> groupIds);
+    List<PermissionDO> selectPermissionsByGroupIds(@Param("groupIds") List<Integer> groupIds);
 
     /**
      * 通过分组id得到所有分组下的权限
@@ -28,7 +28,7 @@ public interface PermissionMapper extends BaseMapper<PermissionDO> {
      * @param groupId 分组id
      * @return 权限
      */
-    List<PermissionDO> selectPermissionsByGroupId(@Param("groupId") Long groupId);
+    List<PermissionDO> selectPermissionsByGroupId(@Param("groupId") Integer groupId);
 
     /**
      * 通过分组ids得到所有分组下的权限
@@ -37,6 +37,6 @@ public interface PermissionMapper extends BaseMapper<PermissionDO> {
      * @param module   权限模块
      * @return 权限
      */
-    List<PermissionDO> selectPermissionsByGroupIdsAndModule(@Param("groupIds") List<Long> groupIds,
+    List<PermissionDO> selectPermissionsByGroupIdsAndModule(@Param("groupIds") List<Integer> groupIds,
                                                             @Param("module") String module);
 }
