+++ src/main/java/vip/gadfly/chakkispring/model/SignListDO.java
@@ -18,12 +18,12 @@ public class SignListDO implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     /**
      * 班级id
      */
-    private Long classId;
+    private Integer classId;
 
     private String name;
